Split DBMetrics in TransactionMetrics, PageCacheMetrics, CheckPointingMetrics, and EntityCountMetrics
